+++ frontend/src/app/features/account-settings/my-account/i18n/fr.json
